+++ src/modules/pool/consts.ts
@@ -1,0 +1,8 @@
+export enum PoolPageNames {
+  PoolDetails = 'PoolDetails',
+}
+
+export enum PoolComponents {
+  AddLiquidityDialog = 'Dialog/AddLiquidity',
+  AddLiquidityForm = 'Form/AddLiquidity',
+}
